refactor(precompile): create crypto submodule for modexp
Move the `modexp` function (both GMP and aurora-engine-modexp
fallback implementations) into a dedicated `crypto.rs` submodule,
separating pure cryptographic logic from the precompile runner
and gas calculation code.

Closes #2782
